server/requirements.txt
Flask==3.1.0
Flask_Cors==5.0.0
Flask_RESTful==0.3.10
psycopg2==2.9.9
python-dotenv==1.1.0
